feat: Connect to Firestore database
This commit connects to the Firestore database by retrieving the instance from the existing Firebase app reference. The db constant exposes the Firestore instance that can be used throughout the code to read and write data from Cloud Firestore. This lays the groundwork for interacting with the database in the Firebase backend